Extracted static methods for retrieving a new location from a given point to a new class
